+++ man-roxygen/return-coefficient.R
@@ -1,0 +1,7 @@
+#' \item \code{coefficients} : A named vector of estimated coefficients
+#' (parameters of the assumed distribution). \strong{Note}: The parameters
+#' are given in location-scale-parameterization. If \code{distribution} is
+#' \code{"weibull"} this is not the parameterization used in
+#' \code{\link[stats:Weibull]{stats::Weibull}}. The parameters of the
+#' shape-scale-parameterization are given under
+#' \code{shape_scale_coefficients}.